#19953 Code style changes
