+++ Cross.Cache.UnitTests/SampleTestDto.cs
@@ -1,0 +1,10 @@
+namespace Cross.Cache.UnitTests;
+
+public class SampleTestDto
+{
+    public int Id { get; set; }
+    
+    public string Code { get; set; }
+
+    public string Description { get; set; }
+}
