+++ Cross.Cache.UnitTests/TestDataWrapper.cs
@@ -1,0 +1,7 @@
+namespace Cross.Cache.UnitTests;
+
+public class TestDataWrapper<T, TExp>
+{
+    public T? Value { get; set; }
+    public TExp? Expected { get; set; }
+}
